app/requirements.txt
flask
gunicorn
google-cloud-storage
qrcode[pil]
